feat: update user and room type imports, enhance user deletion confirmation message, and refactor login and server offline components
