+++ src/app/auth.ts
@@ -73,6 +73,7 @@ export const { handlers, signIn, signOut, auth } = NextAuth({
             }
         })
     ],
+    trustHost: true,
     callbacks: {
         async jwt({ token, user }) {
             if (user) {
