+++ types/user-login.d.ts
@@ -1,0 +1,12 @@
+export enum Role {
+    ADMIN = "ADMIN",
+    AUDITOR = "AUDITOR"
+}
+
+export type User = {
+    id: number
+    username: string
+    password: string,
+    role: Role,
+    is_active: boolean
+}
